Only remove card when greater than record count
On delete there was a crash in
FullCardAssign/FullCardMove because
the cardControls list was shorter than
ViewCardCount. This was in the case where
there were still more records left
than ViewCardCount.

Also, changed the redraw call to FullCardAssgin. The
original call to FullCardMove did not update the UI
cards in either case.